+++ Sem6Task/src/Core/MVC/models/PotokStudyGroup.java
@@ -1,4 +1,4 @@
-package Core.MVC.view.models;
+package Core.MVC.models;
 
 public class PotokStudyGroup {
