+++ Sem6Task/src/Core/MVC/view/TeacherView.java
@@ -1,7 +1,7 @@
 package Core.MVC.view;
 import java.util.List;
 
-import Core.MVC.view.models.Teacher;
+import Core.MVC.models.Teacher;
 
 public class TeacherView implements UserView<Teacher> {
