+++ Sem6Task/src/Core/MVC/view/UserView.java
@@ -1,7 +1,7 @@
 package Core.MVC.view;
 import java.util.List;
 
-import Core.MVC.view.models.User;
+import Core.MVC.models.User;
 
 public interface UserView<T extends User> {
     public void sendOnConsole(List<T> list);
